refactor(memori-esp32): Refactor of BLE and UI for board.
