+++ memori-dev/memori-esp32c3/.harper-dictionary.txt
@@ -1,0 +1,1 @@
+SPI
