--- memori-dev/memori-esp32c3/src/bin/ble_echo.rs
@@ -1,55 +1,0 @@
-#![no_std]
-#![no_main]
-#![deny(
-    clippy::mem_forget,
-    reason = "mem::forget is generally not safe to do with esp_hal types, especially those \
-    holding buffers for the duration of a data transfer."
-)]
-#![deny(clippy::large_stack_frames)]
-
-use embassy_executor::Spawner;
-use esp_backtrace as _;
-
-use esp_hal::clock::CpuClock;
-use esp_hal::timer::timg::TimerGroup;
-use log::info;
-use memori_esp32c3::ble::ble_echo_task;
-use static_cell::StaticCell;
-
-extern crate alloc;
-
-static RADIO: StaticCell<esp_radio::Controller<'static>> = StaticCell::new();
-
-// This creates a default app-descriptor required by the esp-idf bootloader.
-// For more information see: <https://docs.espressif.com/projects/esp-idf/en/stable/esp32/api-reference/system/app_image_format.html#application-description>
-esp_bootloader_esp_idf::esp_app_desc!();
-
-#[allow(
-    clippy::large_stack_frames,
-    reason = "it's not unusual to allocate larger buffers etc. in main"
-)]
-#[esp_rtos::main]
-async fn main(spawner: Spawner) -> () {
-    esp_println::logger::init_logger_from_env();
-
-    let config = esp_hal::Config::default().with_cpu_clock(CpuClock::max());
-    let peripherals = esp_hal::init(config);
-
-    esp_alloc::heap_allocator!(#[esp_hal::ram(reclaimed)] size: 66320);
-    // COEX needs more RAM - so we've added some more
-    esp_alloc::heap_allocator!(size: 170 * 1024);
-
-    let timg0 = TimerGroup::new(peripherals.TIMG0);
-    let sw_interrupt =
-        esp_hal::interrupt::software::SoftwareInterruptControl::new(peripherals.SW_INTERRUPT);
-    esp_rtos::start(timg0.timer0, sw_interrupt.software_interrupt0);
-
-    info!("Embassy initialized!");
-
-    let radio_init: esp_radio::Controller<'_> =
-        esp_radio::init().expect("Failed to initialize Wi-Fi/BLE controller");
-
-    let radio = RADIO.init(radio_init);
-
-    spawner.spawn(ble_echo_task(radio, peripherals.BT)).unwrap();
-}
